powermgmt: Don't force shutdown when externally powered
In rare cases, battery voltage at boot can be below the shutdown
threshold even if a charger is plugged in. This triggers a forced
shutdown and tells you to "RECHARGE!" despite there being plenty
of power available, which is annoying.

Tweak the forced shutdown check so it accounts for external power
sources; if any are present, battery voltage will be ignored.

Change-Id: Id6280b0b666df9eef31c37a03c07c9d6d3f50221